add pgsql index seq view
