+++ .gitignore
@@ -1,6 +1,6 @@
 # Project exclude paths
 
 .idea
-.iml
+*.iml
 /*/*.iml
 target
