+++ data-grip-core-starter/src/main/java/com/javayh/data/grip/core/configuration/AppInfoConditionalOnBean.java
@@ -1,0 +1,27 @@
+package com.javayh.data.grip.core.configuration;
+
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * <p>
+ *
+ * </p>
+ *
+ * @author hai ji
+ * @version 1.0.0
+ * @since 2023-05-30
+ */
+@Configuration
+public class AppInfoConditionalOnBean {
+
+//    @Bean
+//    public DataSource dataSource(DataSourceProperties dataSourceProperties) {
+//        HikariConfig config = new HikariConfig();
+//        config.setJdbcUrl(dataSourceProperties.getUrl());
+//        config.setUsername(dataSourceProperties.getUsername());
+//        config.setPassword(dataSourceProperties.getPassword());
+//        config.setDriverClassName(dataSourceProperties.getDriverClassName());
+//        return new HikariDataSource(config);
+//    }
+
+}
