+++ data-grip-core-starter/src/main/java/com/javayh/data/grip/core/configuration/fiter/DataGripSelector.java
@@ -1,0 +1,23 @@
+package com.javayh.data.grip.core.configuration.fiter;
+
+import org.springframework.context.annotation.ImportSelector;
+import org.springframework.core.type.AnnotationMetadata;
+
+/**
+ * <p>
+ * 选择器
+ * </p>
+ *
+ * @author hai ji
+ * @version 1.0.0
+ * @since 2023-05-29
+ */
+public class DataGripSelector implements ImportSelector {
+
+    @Override
+    public String[] selectImports(AnnotationMetadata annotationMetadata) {
+        return new String[]{
+                "com.javayh.data.grip.core.configuration.properties.DataGripProperties"
+        };
+    }
+}
