+++ data-grip-core-starter/src/main/java/com/javayh/data/grip/core/configuration/properties/common/SelectTablesProperties.java
@@ -1,0 +1,14 @@
+package com.javayh.data.grip.core.configuration.properties.common;
+
+import lombok.Data;
+
+/**
+ * @author haji
+ */
+@Data
+public class SelectTablesProperties {
+
+    private String name;
+    private String ddl;
+
+}
