+++ data-grip-core-starter/src/main/java/com/javayh/data/grip/core/configuration/properties/common/ViewProperties.java
@@ -1,0 +1,20 @@
+package com.javayh.data.grip.core.configuration.properties.common;
+
+import lombok.Data;
+
+/**
+ * <p>
+ *
+ * </p>
+ *
+ * @author hai ji
+ * @version 1.0.0
+ * @since 2023-06-10
+ */
+@Data
+public class ViewProperties {
+
+    private String selectViewName;
+    private String selectViewDdl;
+    private Boolean enable;
+}
