+++ data-grip-core-starter/src/main/java/com/javayh/data/grip/core/package-info.java
@@ -1,0 +1,9 @@
+package com.javayh.data.grip.core;/**
+ * <p>
+ *
+ * </p>
+ *
+ * @author hai ji
+ * @version 1.0.0
+ * @since 2023-06-10
+ */
