+++ data-grip-core-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  com.javayh.data.grip.core.configuration.EnableAutoDataGripConfiguration
