+++ data-grip-exec/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>data-grip</artifactId>
+        <groupId>com.javayh.cdc.data.grip</groupId>
+        <version>1.0.0</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>data-grip-exec</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>com.javayh.cdc.data.grip</groupId>
+            <artifactId>data-grip-core-starter</artifactId>
+        </dependency>
+    </dependencies>
+
+
+</project>
