+++ data-grip-exec/src/main/java/com/javayh/data/grip/exec/DataExecApp.java
@@ -1,0 +1,20 @@
+package com.javayh.data.grip.exec;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * <p>
+ *
+ * </p>
+ *
+ * @author hai ji
+ * @version 1.0.0
+ * @since 2023-06-10
+ */
+@SpringBootApplication
+public class DataExecApp {
+    public static void main(String[] args) {
+        SpringApplication.run(DataExecApp.class, args);
+    }
+}
