+++ doc/httpconfig/generated-requests.http
@@ -1,0 +1,2 @@
+###
+GET http://localhost:8080/datagrip/api/gen/pg
