--- src/main/java/com/javayh/cdc/data/grip/DataGripApp.java
@@ -1,21 +1,0 @@
-package com.javayh.cdc.data.grip;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * <p>
- *
- * </p>
- *
- * @author hai ji
- * @version 1.0.0
- * @since 2023-06-08
- */
-@SpringBootApplication
-public class DataGripApp {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DataGripApp.class, args);
-    }
-}
