--- src/main/java/com/javayh/cdc/data/grip/config/properties/MysqlProperties.java
@@ -1,16 +1,0 @@
-package com.javayh.cdc.data.grip.config.properties;
-
-
-import lombok.Data;
-
-/**
- * @author haiji
- */
-@Data
-public class MysqlProperties {
-
-    private String schema;
-    private SelectTablesProperties selectTables;
-    private FunctionProperties functions;
-
-}
