--- src/main/java/com/javayh/cdc/data/grip/config/properties/OracleProperties.java
@@ -1,15 +1,0 @@
-package com.javayh.cdc.data.grip.config.properties;
-
-import lombok.Data;
-
-/**
- * @author haiji
- */
-@Data
-public class OracleProperties {
-
-    private String namespace;
-    private FunctionProperties functions;
-    private SelectTablesProperties selectTables;
-
-}
