--- src/main/java/com/javayh/cdc/data/grip/config/properties/PostgresProperties.java
@@ -1,15 +1,0 @@
-package com.javayh.cdc.data.grip.config.properties;
-
-import lombok.Data;
-
-/**
- * @author haiji
- */
-@Data
-public class PostgresProperties {
-
-    private String schema;
-    private FunctionProperties functions;
-    private SelectTablesProperties selectTables;
-
-}
