--- src/main/java/com/javayh/cdc/data/grip/config/properties/SelectTablesProperties.java
@@ -1,14 +1,0 @@
-package com.javayh.cdc.data.grip.config.properties;
-
-import lombok.Data;
-
-/**
- * @author haji
- */
-@Data
-public class SelectTablesProperties {
-
-    private String name;
-    private String ddl;
-
-}
