--- src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,23 +1,0 @@
-{
-  "properties": [
-    {
-      "name": "data.grip.postgres.create-function",
-      "type": "java.lang.String",
-      "description": "Description for data.grip.postgres.create-function."
-    },
-    {
-      "name": "data.grip.postgres.schema",
-      "type": "java.lang.String",
-      "description": "Description for data.grip.postgres.schema."
-    },
-    {
-      "name": "data.grip.postgres.select-tables.name",
-      "type": "java.lang.String",
-      "description": "Description for data.grip.postgres.select-tables.name."
-    },
-    {
-      "name": "data.grip.postgres.select-tables.ddl",
-      "type": "java.lang.String",
-      "description": "Description for data.grip.postgres.select-tables.ddl."
-    }
-  ] }
