--- src/main/resources/application-datagrip.yml
@@ -1,17 +1,0 @@
-data:
-  grip:
-    postgres:
-      schema: public
-      functions:
-        function-name: generate_create_table_statement
-        ddl: datasource/pgsql/carete_funtion.sql
-        ddl-enable: false
-      select-tables:
-        name:
-          SELECT a.oid, a.relname AS name,b.description AS comment
-          FROM pg_class a
-          LEFT OUTER JOIN pg_description b ON b.objsubid=0 AND a.oid = b.objoid
-          WHERE a.relnamespace = (SELECT oid FROM pg_namespace WHERE nspname= 'public')
-          AND a.relkind='r'
-          ORDER BY a.relname;
-
